Create a new parent pom hierarchy to fit with the applications-poms:
   helidon-parent -> helidon-bom -> helidon-3rdparty -> helidon-project
                                                     -> applications-project

The BOM pom is now inherited by helidon-project, thus the requirement of ${project.version} for
Helidon modules is now dropped. This also helps enforcing that the BOM pom contains an entry
for all modules.

Using <relativePath>../../../applications/se/pom.xml</relativePath> in all examples in order
to build locally with a single Maven reactor. This shall be refactored to <relativePath/> when
the example are moved out of the main repository. It is a temporary trade-off.