Add helidon-config-yaml-mp as a dependency in helidon-microprofile-config. (#4379)
helidon-config-yaml-mp was recently decoupled from helidon-config-mp which is part of helidon-microprofile-config and as a consequence, is not a part of that config bundle anymore. This work was done to add it back and avoid any effects of incompatibility.

(cherry picked from commit aca9bf76f046fd94f292c8a6d61ae79c65f72031)