Fix SLF4J binding for annotation processors
Add slf4j-api as a dependency to inject/tools.
This ensures when the compiler plugin resolves the processor transitive dependencies the resolved version of slf4j-api is the managed version.
Otherwise we may end-up with different versions between slf4j-api and slf4j-jdk14 and a failed binding with the warning.

Fixes #7129