Avoid implementing the OCI Monitoring interface and instead use Mockito to mock it. (#8474)
The goal of this test fix is to remove implementing Monitoring interface as a way to mock Monitoring client, as it is causing  updates to be made if that interface has changed during an upgrade to OCI SDK is made. Instead, the OCIExtension is now mocked and inside of it, a mocked Monitoring bean using Proxy is injected. With this approach, there is no need to  implement all the methods of Monitoring interface, but only mock those that are needed for the test.